Better separate Inequality logic, components, and related logic that is used in other places
