+++ src/app/services/questions.ts
@@ -87,6 +87,67 @@ export function generateQuestionTitle(doc : ContentDTO | ContentSummaryDTO) {
     return title;
 }
 
+// Inequality specific functions
+
+export function sanitiseInequalityState(state: any) {
+    const saneState = JSON.parse(JSON.stringify(state));
+    if (saneState.result?.tex) {
+        saneState.result.tex = saneState.result.tex.split('').map((l: string) => REVERSE_GREEK_LETTERS_MAP[l] ? '\\' + REVERSE_GREEK_LETTERS_MAP[l] : l).join('');
+    }
+    if (saneState.result?.python) {
+        saneState.result.python = saneState.result.python.split('').map((l: string) => REVERSE_GREEK_LETTERS_MAP[l] || l).join('');
+    }
+    if (saneState.result?.uniqueSymbols) {
+        saneState.result.uniqueSymbols = saneState.result.uniqueSymbols.split('').map((l: string) => REVERSE_GREEK_LETTERS_MAP[l] || l).join('');
+    }
+    if (saneState.symbols) {
+        for (const symbol of saneState.symbols) {
+            if (symbol.expression.latex) {
+                symbol.expression.latex = symbol.expression.latex.split('').map((l: string) => REVERSE_GREEK_LETTERS_MAP[l] ? '\\' + REVERSE_GREEK_LETTERS_MAP[l] : l).join('');
+            }
+            if (symbol.expression.python) {
+                symbol.expression.python = symbol.expression.python.split('').map((l: string) => REVERSE_GREEK_LETTERS_MAP[l] || l).join('');
+            }
+        }
+    }
+    return saneState;
+}
+
+export const parsePseudoSymbolicAvailableSymbols = (availableSymbols?: string[]) => {
+    if (!availableSymbols) return;
+    const theseSymbols = availableSymbols.slice(0).map(s => s.trim());
+    let i = 0;
+    while (i < theseSymbols.length) {
+        if (theseSymbols[i] === '_trigs') {
+            theseSymbols.splice(i, 1, 'cos()', 'sin()', 'tan()');
+            i += 3;
+        } else if (theseSymbols[i] === '_1/trigs') {
+            theseSymbols.splice(i, 1, 'cosec()', 'sec()', 'cot()');
+            i += 3;
+        } else if (theseSymbols[i] === '_inv_trigs') {
+            theseSymbols.splice(i, 1, 'arccos()', 'arcsin()', 'arctan()');
+            i += 3;
+        } else if (theseSymbols[i] === '_inv_1/trigs') {
+            theseSymbols.splice(i, 1, 'arccosec()', 'arcsec()', 'arccot()');
+            i += 3;
+        } else if (theseSymbols[i] === '_hyp_trigs') {
+            theseSymbols.splice(i, 1, 'cosh()', 'sinh()', 'tanh()', 'cosech()', 'sech()', 'coth()');
+            i += 6;
+        } else if (theseSymbols[i] === '_inv_hyp_trigs') {
+            theseSymbols.splice(i, 1, 'arccosh()', 'arcsinh()', 'arctanh()', 'arccosech()', 'arcsech()', 'arccoth()');
+            i += 6;
+        } else if (theseSymbols[i] === '_logs') {
+            theseSymbols.splice(i, 1, 'log()', 'ln()');
+            i += 2;
+        } else if (theseSymbols[i] === '_no_alphabet') {
+            theseSymbols.splice(i, 1);
+        } else {
+            i += 1;
+        }
+    }
+    return theseSymbols;
+};
+
 /**
  * Essentially a useState for the current question attempt - used in all question components.
  * @param questionId  id of the question to return the current attempt of
